fix(index): enforce chunk budget and ship retrieval gate
