+++ .env.example
@@ -255,19 +255,23 @@ EMBEDDING_DIMENSIONS=384
 # content, and budgets picked by measuring a ratio have been refuted three times
 # by the next content class (384 by Korean prose, 352 by minified JSON, 284 by
 # this repository's own minified JS). cognee also emits an over-budget chunk
-# verbatim when a single word does not fit, so no setting here caps chunk size
-# on its own. Citadel reports every real overflow at the embed boundary instead;
+# verbatim when a single word does not fit. Citadel replays the pinned Cognee
+# chunker before normal ingest and rejects a final chunk whose reported or exact
+# text size exceeds this budget. Existing rows still need an approved repair.
+# The embed-boundary detector remains telemetry for model-window overflow;
 # grep the logs for "embed window exceeded" and lower this when one appears.
 # Leave unset to take the value above. Lowering it raises the chunk count, which
 # costs cognify time and rows, so it is worth a measurement rather than a guess.
 CITADEL_CHUNK_BUDGET_TOKENS=256
 # Report chunks that exceeded the embedder's window. On by default; a detector
-# that is off is indistinguishable from one with nothing to report.
+# that is off is indistinguishable from one with nothing to report. This is
+# separate from the pre-storage chunk-budget check above.
 CITADEL_EMBED_WINDOW_DETECTOR=true
 # Refuse a document containing one unbroken word larger than the budget. cognee
 # breaks words on a space and on sentence endings only, so one minified line is
-# one word: such a document either fails the whole dataset's cognify pass or is
-# stored truncated. The policy is refuse and record; nothing here edits content.
+# one word: such a document would either fail the whole dataset's cognify pass
+# or be stored truncated without the pre-storage check. The policy is refuse and
+# record; nothing here edits content.
 CITADEL_UNCHUNKABLE_GUARD=true
 
 # Railway/Postgres examples. Bind DATABASE_URL from the Railway Postgres service.
